+++ package.json
@@ -11,8 +11,8 @@
     "demo.android": "npm run demo.reset && cd demo && tns debug android --no-hmr",
     "demo.ios": "npm run demo.reset && cd demo && tns debug ios --no-hmr",
     "demo.reset": "rimraf demo/platforms src/platforms/android/nativescript_audioplayer.aar",
-    "format:check": "prettier-check --config ./.prettierrc.json 'src/*.ts' 'demo/src/**/*.{ts,html,css,scss}'",
-    "format": "prettier --write --config ./.prettierrc.json 'src/*.ts' 'demo/src/**/*.{ts,html,css,scss}'",
+    "format:check": "prettier-check --config ./.prettierrc.json 'src/*.ts' 'src/native-definitions/**/*.ts' 'demo/src/**/*.{ts,html,css,scss}'",
+    "format": "prettier --write --config ./.prettierrc.json 'src/*.ts' 'src/native-definitions/**/*.ts' 'demo/src/**/*.{ts,html,css,scss}'",
     "gen.typings.ios": "cd demo && TNS_TYPESCRIPT_DECLARATIONS_PATH=\"$(pwd)/../gen-typings\" tns build ios",
     "lint": "npm-run-all format:check tslint",
     "test.android": "(cd src && npm i && npx tsc) && npm run tslint && cd demo && tns build android && tns test android --justlaunch",
