+++ src/audioplayer-common.ts
@@ -37,7 +37,33 @@ export abstract class CommonAudioPlayer {
     return this.seekIntervalSeconds;
   }
 
-  public abstract setSleepTimer(milliseconds: number);
+  public setSleepTimer(milliseconds: number) {
+    if (trace.isEnabled()) {
+      trace.write(`${this.cls}.setSleepTimer(${milliseconds})`, notaAudioCategory);
+    }
+
+    this.cancelSleepTimer();
+
+    const countdownTick = 1000;
+    this._sleepTimerMillisecondsLeft = milliseconds;
+    this._sleepTimer = setInterval(() => {
+      if (!this._sleepTimerPaused && this.isPlaying()) {
+        this._sleepTimerMillisecondsLeft = Math.max(this._sleepTimerMillisecondsLeft - countdownTick, 0);
+        this._onPlaybackEvent(PlaybackEvent.SleepTimerChanged);
+      }
+
+      if (this._sleepTimerMillisecondsLeft === 0) {
+        // Fade out volume and pause if not already paused.
+        if (this.isPlaying()) {
+          this.onSleepTimerExpired();
+        }
+        clearInterval(this._sleepTimer);
+        this._sleepTimer = undefined;
+      }
+    }, countdownTick);
+
+    this._onPlaybackEvent(PlaybackEvent.SleepTimerChanged);
+  }
 
   public getSleepTimerRemaining(): number {
     return this._sleepTimerMillisecondsLeft;
@@ -48,12 +74,14 @@ export abstract class CommonAudioPlayer {
       trace.write(`${this.cls}.cancelSleepTimer()`, notaAudioCategory);
     }
 
-    if (this._sleepTimer !== undefined) {
-      clearInterval(this._sleepTimer);
-      this._sleepTimer = undefined;
-      this._sleepTimerMillisecondsLeft = 0;
-      this._onPlaybackEvent(PlaybackEvent.SleepTimerChanged);
+    if (this._sleepTimer === undefined) {
+      return;
     }
+
+    clearInterval(this._sleepTimer);
+    this._sleepTimer = undefined;
+    this._sleepTimerMillisecondsLeft = 0;
+    this._onPlaybackEvent(PlaybackEvent.SleepTimerChanged);
   }
 
   protected _sleepTimer: number;
@@ -61,23 +89,27 @@ export abstract class CommonAudioPlayer {
   protected _sleepTimerMillisecondsLeft = 0;
 
   public pauseSleepTimer() {
-    if (this._sleepTimer !== undefined) {
-      if (trace.isEnabled()) {
-        trace.write(`${this.cls}.pauseSleepTimer()`, notaAudioCategory);
-      }
-
-      this._sleepTimerPaused = true;
+    if (this._sleepTimer === undefined) {
+      return;
     }
+
+    if (trace.isEnabled()) {
+      trace.write(`${this.cls}.pauseSleepTimer()`, notaAudioCategory);
+    }
+
+    this._sleepTimerPaused = true;
   }
 
   public resumeSleepTimer() {
-    if (this._sleepTimer !== undefined) {
-      if (trace.isEnabled()) {
-        trace.write(`${this.cls}.resumeSleepTimer()`, notaAudioCategory);
-      }
-
-      this._sleepTimerPaused = false;
+    if (this._sleepTimer === undefined) {
+      return;
     }
+
+    if (trace.isEnabled()) {
+      trace.write(`${this.cls}.resumeSleepTimer()`, notaAudioCategory);
+    }
+
+    this._sleepTimerPaused = false;
   }
 
   public abstract destroy();
@@ -86,24 +118,30 @@ export abstract class CommonAudioPlayer {
     this.preparePlaylist(playlist);
     if (startIndex !== undefined && startOffset) {
       this.skipToPlaylistIndexAndOffset(startIndex, startOffset);
-    } else if (startIndex) {
+      return;
+    }
+
+    if (startIndex) {
       this.skipToPlaylistIndex(startIndex);
+      return;
     }
   }
 
   public skipToPlaylistIndexAndOffset(playlistIndex: number, offset: number): void {
     if (this.getCurrentPlaylistIndex() === playlistIndex) {
       this.seekTo(offset);
-    } else {
-      if (offset > 0) {
-        if (trace.isEnabled()) {
-          trace.write(`Set queuedSeek to ${offset}`, notaAudioCategory);
-        }
-        this._queuedSeekTo = offset;
+      return;
+    }
+
+    if (offset > 0) {
+      if (trace.isEnabled()) {
+        trace.write(`Set queuedSeek to ${offset}`, notaAudioCategory);
       }
 
-      this.skipToPlaylistIndex(playlistIndex);
+      this._queuedSeekTo = offset;
     }
+
+    this.skipToPlaylistIndex(playlistIndex);
   }
 
   public seekRelative(relativeOffset: number): void {
@@ -119,8 +157,35 @@ export abstract class CommonAudioPlayer {
   }
 
   public _onPlaybackEvent(evt: PlaybackEvent, arg?: any) {
-    if (this._listener) {
-      this._listener.onPlaybackEvent(evt, arg);
+    try {
+      switch (evt) {
+        case PlaybackEvent.Playing: {
+          this.resumeSleepTimer();
+          break;
+        }
+        case PlaybackEvent.Stopped: {
+          this.cancelSleepTimer();
+          break;
+        }
+        case PlaybackEvent.Paused:
+        case PlaybackEvent.WaitingForNetwork:
+        case PlaybackEvent.EncounteredError: {
+          this.pauseSleepTimer();
+          break;
+        }
+      }
+    } catch (err) {
+      trace.write(`${this.cls}._onPlaybackEvent(${evt}) - handle sleep timer failed. ${err}`, notaAudioCategory, trace.messageType.error);
     }
+
+    if (!this._listener) {
+      return;
+    }
+
+    this._listener.onPlaybackEvent(evt, arg);
+  }
+
+  protected onSleepTimerExpired() {
+    this.pause();
   }
 }
