+++ src/audioplayer.d.ts
@@ -29,4 +29,4 @@ export class TNSAudioPlayer {
   _onPlaybackEvent(evt: PlaybackEvent, args?: any);
 }
 
-export * from './audioplayer.types'
+export * from './audioplayer.types';
