+++ src/package.json
@@ -31,8 +31,8 @@
   "readmeFilename": "README.md",
   "devDependencies": {
     "tns-core-modules": "^6.0.0",
-    "typescript": "3.4.5",
-    "tslint": "^5.18.0"
+    "tslint": "^5.18.0",
+    "typescript": "^3.4.5"
   },
   "dependencies": {}
 }
