+++ src/tsconfig.json
@@ -18,7 +18,8 @@
         "noImplicitAny": false,
         "noImplicitReturns": true,
         "noImplicitUseStrict": false,
-        "noFallthroughCasesInSwitch": true
+        "noFallthroughCasesInSwitch": true,
+        "incremental": true
     },
     "exclude": [
         "node_modules",
